+++ DiFfRG/include/DiFfRG/discretization/grid/stack_coordinates.hh
@@ -9,7 +9,7 @@ namespace DiFfRG
   {
   public:
     using ctype = Idx;
-    static constexpr uint dim = 1;
+    static constexpr size_t dim = 1;
 
     IndexStack(Idx start, Idx stop) : start(start), stop(stop), m_size(stop - start)
     {
@@ -36,19 +36,19 @@ namespace DiFfRG
      */
     Idx KOKKOS_FORCEINLINE_FUNCTION backward(const Idx y) const { return y - start; }
 
-    uint size() const { return m_size; }
+    size_t size() const { return m_size; }
 
     const Idx start, stop;
 
   private:
-    const uint m_size;
+    const size_t m_size;
   };
 
   template <typename Idx = int, typename NT = double> class BosonicMatsubaraValues
   {
   public:
     using ctype = NT;
-    static constexpr uint dim = 1;
+    static constexpr size_t dim = 1;
 
     BosonicMatsubaraValues(Idx start, Idx stop, NT T) : start(start), stop(stop), T(T), m_size(stop - start)
     {
@@ -79,20 +79,20 @@ namespace DiFfRG
      */
     Idx KOKKOS_FORCEINLINE_FUNCTION backward(const NT &y) const { return Idx(std::round(y / (2. * M_PI * T))) - start; }
 
-    uint size() const { return m_size; }
+    size_t size() const { return m_size; }
 
     const Idx start, stop;
     const NT T;
 
   private:
-    const uint m_size;
+    const size_t m_size;
   };
 
   template <typename Idx = int, typename NT = double> class FermionicMatsubaraValues
   {
   public:
     using ctype = NT;
-    static constexpr uint dim = 1;
+    static constexpr size_t dim = 1;
 
     FermionicMatsubaraValues(Idx start, Idx stop, NT T) : start(start), stop(stop), T(T), m_size(stop - start)
     {
@@ -126,12 +126,12 @@ namespace DiFfRG
       return Idx(std::round((y - M_PI * T) / (2. * M_PI * T))) - start;
     }
 
-    uint size() const { return m_size; }
+    size_t size() const { return m_size; }
 
     const Idx start, stop;
     const NT T;
 
   private:
-    const uint m_size;
+    const size_t m_size;
   };
 } // namespace DiFfRG
